sales admin panel in progress

+++ database/migrations/2022_06_10_090705_t_sales_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('sales', function (Blueprint $table) {
             $table->id();
-            $table->integer('ticket_number');
+            $table->bigInteger('ticket_number');
             $table->date('date_emission');
             $table->time('time_emission');
             $table->integer('payment_method_id');
